Update Product, DataBase and DummyDataBase: change Long type of price to Integer.

+++ src/main/java/ru/akirakozov/sd/refactoring/databse/DataBase.java
@@ -6,7 +6,7 @@ public interface DataBase {
     public void             addProduct(Product product);
     public List<Product>    selectAllProducts();
     public Integer          getProductsCount();
-    public Long             getProductPricesSum();
+    public Integer             getProductPricesSum();
     public Product          getProductWithMinPrice();
     public Product          getProductWithMaxPrice();
 }
